Refactor output formatting. Add support for 8/8/2018 Endpoints (#86)
* Refactor output formatting

* Remove extraneous files

* Add output formatting for Stocks methods. Change to DatetimeIndex

* Add back moved functions as deprecated
