+++ .gitignore
@@ -109,3 +109,4 @@ venv.bak/
 
 # mypy
 .mypy_cache/
+sftp-config\.json
