+++ .travis.yml
@@ -20,7 +20,7 @@ install:
   - pip install codecov
   - if [[ $TRAVIS_PYTHON_VERSION == 3.6 ]]; then
       pip install doctr sphinx sphinx_rtd_theme ipython matplotlib;
-      pip install sphinxcontrib-napoleon;
+      pip install sphinxcontrib-napoleon requests-cache;
     fi
   - python setup.py install
