+++ iexfinance/base.py
@@ -15,21 +15,11 @@ class _IEXBase(object):
     Base class for retrieving equities information from the IEX Finance API.
     Inherited by Stock and Market Readers, and conducts query operations
     including preparing and executing queries from the API.
-
-    Attributes
-    ----------
-    retry_count: int, default 3, optional
-        Desired number of retries if a request fails
-    pause: float, default 0.5, optional
-        Pause time between retry attempts
-    session: requests_cache.session, default None, optional
-        A cached requests-cache session
     """
     # Base URL
     _IEX_API_URL = "https://api.iextrading.com/1.0/"
 
-    def __init__(self, retry_count=3, pause=0.5, session=None,
-                 json_parse_int=None, json_parse_float=None):
+    def __init__(self, **kwargs):
         """ Initialize the class
 
         Parameters
@@ -44,12 +34,15 @@ class _IEXBase(object):
             Desired integer parsing datatype
         json_parse_float: datatype, default float, optional
             Desired floating point parsing datatype
+        output_format: str, default "json", optional
+            Desired output format (json or pandas DataFrame)
         """
-        self.retry_count = retry_count
-        self.pause = pause
-        self.session = _init_session(session)
-        self.json_parse_int = json_parse_int
-        self.json_parse_float = json_parse_float
+        self.retry_count = kwargs.get("retry_count", 3)
+        self.pause = kwargs.get("pause", 0.5)
+        self.session = _init_session(kwargs.get("session"))
+        self.json_parse_int = kwargs.get("json_parse_int")
+        self.json_parse_float = kwargs.get("json_parse_float")
+        self.output_format = kwargs.get("output_format", 'json')
 
     @property
     def params(self):
@@ -111,6 +104,7 @@ class _IEXBase(object):
         """
         for i in range(self.retry_count+1):
             response = self.session.get(url=url, params=self.params)
+            print("URL: %s" % response.request.url)
             if response.status_code == requests.codes.ok:
                 return self._validate_response(response)
             time.sleep(self.pause)
@@ -126,7 +120,7 @@ class _IEXBase(object):
         """
         return "%s%s" % (self._IEX_API_URL, self.url)
 
-    def fetch(self):
+    def fetch(self, fmt_p=None, fmt_j=None):
         """Fetches latest data
 
         Prepares the query URL based on self.params and executes the request
@@ -137,4 +131,22 @@ class _IEXBase(object):
             A response object
         """
         url = self._prepare_query()
-        return self._execute_iex_query(url)
+        data = self._execute_iex_query(url)
+        return self._output_format(data, fmt_j=fmt_j, fmt_p=fmt_p)
+
+    def _convert_output(self, out):
+        import pandas as pd
+        return pd.DataFrame(out)
+
+    def _output_format(self, out, fmt_j=None, fmt_p=None):
+        """
+        Output formatting handler
+        """
+        if self.output_format == 'pandas':
+            if fmt_p:
+                return fmt_p(out)
+            else:
+                return self._convert_output(out)
+        if fmt_j:
+            return fmt_j(out)
+        return out
