+++ iexfinance/stocks/collections.py
@@ -1,0 +1,34 @@
+import pandas as pd
+
+from iexfinance.base import _IEXBase
+
+# Data provided for free by IEX
+# See https://iextrading.com/api-exhibit-a/ for additional information
+# and conditions of use
+
+
+class CollectionsReader(_IEXBase):
+    """
+    Class for retrieving data from the Collections endpoint
+    """
+    _COLLECTION_TYPES = ["tag", "sector", "list"]
+
+    def __init__(self, collection_name, collection_type="tag", **kwargs):
+        self.collection_name = collection_name
+        self.collection_type = collection_type
+        if self.collection_type not in self._COLLECTION_TYPES:
+            raise ValueError("Please select a valid collection type.")
+        super(CollectionsReader, self).__init__(**kwargs)
+
+    @property
+    def url(self):
+        return "stock/market/collection/%s" % self.collection_type
+
+    @property
+    def params(self):
+        return {
+            "collectionName": self.collection_name
+        }
+
+    def _convert_output(self, out):
+        return pd.DataFrame(out).set_index("symbol")
