+++ iexfinance/stocks/todayearnings.py
@@ -1,0 +1,11 @@
+from iexfinance.base import _IEXBase
+
+
+class EarningsReader(_IEXBase):
+
+    @property
+    def url(self):
+        return "stock/market/today-earnings"
+
+    def _convert_output(self, out):
+        return out
