+++ requirements-test.txt
@@ -1,3 +1,4 @@
 flake8
 pytest
 pytest-runner
+six
